added shared folder and moved things around
