+++ python3/agents/__init__.py
@@ -1,0 +1,2 @@
+from .vm_totoro_agent import *
+from .totoro_agent import *
