+++ python3/agents/shared/strategies/bomb_strategy.py
@@ -1,4 +1,3 @@
-import random
 from typing import List
 
 from . import strategy
