+++ python3/agents/shared/strategies/detonate_strategy.py
