+++ python3/agents/shared/trackers/__init__.py
@@ -1,0 +1,4 @@
+from .bomb_tracker import BombTracker
+from .enemy_tracker import EnemyTracker
+from .pickup_tracker import PickupTracker
+from .map_tracker import MapTracker
