+++ python3/agents/totoro_agent/brain.py
@@ -2,10 +2,7 @@
 Tracks the game state and decides on next strategy to execute
 """
 
-from .bomb_tracker import BombTracker
-from .enemy_tracker import EnemyTracker
-from .map_tracker import MapTracker
-from .pickup_tracker import PickupTracker
+from ..shared.trackers import BombTracker, EnemyTracker, MapTracker, PickupTracker
 
 
 class Brain:
@@ -42,34 +39,33 @@ class Brain:
         # should probably refine this to check opponent vulnerability and trappable
         # Uncomment after done with basic_avoid
 
-       # If you're in the blast tiles, stop being in them loser.
-        #print(game_state['tick'], "player pos", game_state['player_pos'], "enemy bombs", game_state['enemy_active_bombs'], "\n Game State hazard zones:", game_state['hazard_zones'], game_state['player_pos'] in game_state['hazard_zones'])
-
-
-
-
+        # If you're in the blast tiles, stop being in them loser.
+        # print(game_state['tick'], "player pos", game_state['player_pos'], "enemy bombs", game_state['enemy_active_bombs'], "\n Game State hazard zones:", game_state['hazard_zones'], game_state['player_pos'] in game_state['hazard_zones'])
 
         # If you're in the blast tiles, do RETREAT
-        #print(game_state['enemy_immediate_trapped'], "in brain")
-        if (game_state['player_pos'] in game_state['all_hazard_zones'] or (game_state['player_on_bomb']) and not game_state['player_is_invulnerable']):
+        # print(game_state['enemy_immediate_trapped'], "in brain")
+        if (game_state['player_pos'] in game_state['all_hazard_zones'] or (game_state['player_on_bomb']) and not
+        game_state['player_is_invulnerable']):
             print('HOLY RUN FOR YOUR LIFE YOU ARE GONNA GET RAILED')
-            return 'basic_avoid' # Basic avoid vs retreat. Retreat value based, basic avoid is coded.
-
+            return 'basic_avoid'  # Basic avoid vs retreat. Retreat value based, basic avoid is coded.
 
         # if enemy in detonation zone and we aren't + if they aren't invulnerable (we'll move out of the way via basic_avoid's top priority)
-        if not game_state['enemy_is_invulnerable'] and (game_state['enemy_pos'] in game_state['detonation_zones'] and (game_state['player_pos'] not in game_state['detonation_zones'])):
+        if not game_state['enemy_is_invulnerable'] and (game_state['enemy_pos'] in game_state['detonation_zones'] and (
+                game_state['player_pos'] not in game_state['detonation_zones'])):
             print('KABOOOM!! DETONATION TIME!')
             return 'detonate'
 
         # Hard-coding immediate trap (can put in a strategy later)
         ## Check if enemy is trapped: ->check if player can place a bomb that attacks enemy: -> do it.
-        elif game_state['enemy_immediate_trapped'] and (game_state['player_inv_bombs'] > 0 and not game_state['enemy_near_bomb']): # Immediate trapped also takes into account whether the player is there.
-            
-            print(game_state['enemy_immediate_trapped'], game_state['player_inv_bombs'] > 0 and not game_state['enemy_near_bomb'])
+        elif game_state['enemy_immediate_trapped'] and (game_state['player_inv_bombs'] > 0 and not game_state[
+            'enemy_near_bomb']):  # Immediate trapped also takes into account whether the player is there.
+
+            print(game_state['enemy_immediate_trapped'],
+                  game_state['player_inv_bombs'] > 0 and not game_state['enemy_near_bomb'])
             print("I think the enemy is trapped so I'm placing a bomb right now!!", game_state['tick'])
-            #place bomb 
-            return "bomb" # literally just fucking bomb them??
-            
+            # place bomb
+            return "bomb"  # literally just fucking bomb them??
+
         # Pickup if ammo, stalk if none on map.
         elif len(game_state['pickup_list']) != 0:  # "Any pickups on the map?"
             print('me gusta I smell some pickups')
