+++ python3/agents/vm_totoro_agent/my_agent.py
@@ -1,4 +1,4 @@
-from .strategies import RandomStrategy, RetreatStrategy, StalkStrategy, PickupStrategy, AdvKillStrategy, BasicAvoidStrategy, DetonateStrategy, BombStrategy, BlockDestroyingStrategy
+from ..shared.strategies import RandomStrategy, RetreatStrategy, StalkStrategy, PickupStrategy, AdvKillStrategy, DetonateStrategy, BlockDestroyingStrategy
 from .brain import Brain
 
 
@@ -10,11 +10,9 @@ class Agent:
             'retreat': RetreatStrategy(),
             'pickup': PickupStrategy(),
             'stalk': StalkStrategy(),
-            'basic_avoid': BasicAvoidStrategy(),
             'kill': AdvKillStrategy(),
-            'detonate': DetonateStrategy(),
-            'bomb': BombStrategy(),
             'block_destroy': BlockDestroyingStrategy(),
+            'detonate': DetonateStrategy()
         }
         self.action_queue = []
         self.prev_tick = -1
