--- python3/totoro_agent/brain/__init__.py
@@ -1,5 +1,0 @@
-"""
-Defines the logic for decision making
-"""
-
-from .brain import Brain
