--- python3/vm_totoro_agent/strategies/__init__.py
@@ -1,11 +1,0 @@
-"""
-Contains different strategies used in the game
-"""
-
-from .random_strategy import RandomStrategy
-from .retreat_strategy import RetreatStrategy
-from .pickup_strategy import PickupStrategy
-from .stalk_strategy import StalkStrategy
-from .adv_kill_strategy import AdvKillStrategy
-from .block_destroying import BlockDestroyingStrategy
-from .detonate_strategy import DetonateStrategy
