--- python3/vm_totoro_agent/strategies/strategy.py
@@ -1,13 +1,0 @@
-"""
-Interface for defining strategy pattern.
-"""
-from typing import List
-
-
-class Strategy:
-
-    def execute(self, game_state: object) -> List[str]:
-        """
-        Execute the strategy
-        """
-        pass
