--- python3/vm_totoro_agent/utils/__init__.py
@@ -1,7 +1,0 @@
-"""
-Utility helper functions and constants
-"""
-
-
-from . import constants
-from . import util_functions
