--- python3/vm_totoro_agent/utils/structures/__init__.py
@@ -1,1 +1,0 @@
-from .graph_node import Node
